refactor : 쉐어 통합 테스트 관련 Utils 클래스 패키지 분리
